+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 
 #### **Languages:**
 - The binding attribute "data-syntax-language" can now be set to "unknown", which will force the layout to still be drawn, but no highlighting will be applied.
+- Added a new JSON property called "keywordsCasing", which states what casing should be used for the keywords when rendered (optional, defaults to "initial", accepts "uppercase" and "lowercase").
 
 #### **Binding Settings:**
 - Added a new setting called "padLineNumbers", which states if the line numbers should be padded (defaults to false).